+++ locksmith/__tests__/worker/notify.test.ts
@@ -1,4 +1,4 @@
-import { handler } from '../../__mocks__/websub/subscriber'
+import { handler } from '../../__mocks__/worker/subscriber'
 import { notify } from '../../src/worker/helpers'
 import { Hook, HookEvent } from '../../src/models'
 import { vi } from 'vitest'
